🐳 chore: 페이지 폴더구조 세팅
