--- src/app/admin/posts/page.tsx
@@ -1,4 +1,0 @@
-//전체 게시글 목록
-export default function AdminPostsPage() {
-  return <h2>전체 게시글 목록</h2>;
-}
