+++ src/app/auth/login/page.tsx
@@ -1,4 +1,4 @@
 // 로그인
 export default function LoginPage() {
-  return <h2>로그인</h2>;
+  return <h2>로그인페이지</h2>;
 }
