+++ src/app/auth/signup/page.tsx
@@ -1,0 +1,4 @@
+// 회원가입
+export default function SingUpPage() {
+  return <h2>회원가입페이지</h2>;
+}
