+++ src/app/feed/page.tsx
@@ -1,0 +1,4 @@
+// 살펴보기 게시판 목록
+export default function FeedPage() {
+  return <h2>살펴보기 게시판</h2>;
+}
