+++ src/app/mypage/bookmark/page.tsx
@@ -1,0 +1,4 @@
+// 마이페이지/북마크
+export default function MypageBookmarkPage() {
+  return <h2>마이페이지/북마크</h2>;
+}
