+++ src/app/mypage/page.tsx
@@ -1,0 +1,4 @@
+// 마이페이지/리뷰
+export default function MypageReviewPage() {
+  return <h2>마이페이지/리뷰</h2>;
+}
