+++ src/app/photomap/page.tsx
@@ -1,0 +1,4 @@
+// 지도만들기 게시판 목록
+export default function FeedPage() {
+  return <h2>지도만들기 게시판</h2>;
+}
