+++ src/app/plan/[id]/page.tsx
