+++ src/app/plan/page.tsx
@@ -1,0 +1,4 @@
+// 여행계획
+export default function TravelPage() {
+  return <h2>여행계획</h2>;
+}
