--- src/app/post/page.tsx
@@ -1,4 +1,0 @@
-// 여행지상세조회
-export default function TravelPage() {
-  return <h2>여행지후기조회</h2>;
-}
