--- src/app/profile/page.tsx
@@ -1,4 +1,0 @@
-// 마이페이지
-export default function Mypage() {
-  return <h2>마이페이지</h2>;
-}
