+++ src/app/review/[id]/edit/page.tsx
@@ -1,0 +1,4 @@
+// 자유게시판 목록
+export default function FreeEditPage() {
+  return <h2>자유게시판</h2>;
+}
