+++ src/app/review/new/page.tsx
@@ -1,0 +1,4 @@
+// 자유게시판 목록
+export default function FreeAddPage() {
+  return <h2>자유게시판</h2>;
+}
