+++ src/app/review/page.tsx
@@ -1,0 +1,4 @@
+// 후기 작성
+export default function ReviewPage() {
+  return <h2>후기작성</h2>;
+}
