--- src/app/signup/page.tsx
@@ -1,4 +1,0 @@
-// 회원가입
-export default function LoginPage() {
-  return <h2>회원가입</h2>;
-}
